pyproject.toml
[tool.poetry]
name = "aiodeta"
version = "0.1.0"
description = "Unofficial Deta client"
authors = ["leits <leits.dev@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
aiohttp = "^3.7.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-asyncio = "^0.15.1"
black = "^21.7b0"
isort = "^5.9.3"
coverage = "^5.5"
tox = "^3.24.1"
python-dotenv = "^0.19.0"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
